52 проц покрытия тестами
