--- tests/test_transaction.py
@@ -1,5 +1,0 @@
-from src.transaction import summ
-
-
-def test_summ():
-    assert summ(4, 3) == 7
